New file for TopicTableViewCell + minor fixes
